추첨 확률 추가, 라이트모드 고정, 오류 수정

+++ modules/윤년 생일 문제 확률.py
@@ -46,7 +46,7 @@ if mode:
     st.plotly_chart(line(chart_data, float(f)), use_container_width=True)
     st.dataframe(cut10(chart_data, number))
 
-    st.write(number, "개의 그룹에 그룹당 ", n, "명의 사람들이 있을 때 그룹 안에서 생일이 같은 사람이 생길 확률은 ", l[-1][0], "이다. ")
+    st.write(number, "개의 그룹에 그룹당 ", n, "명의 사람들이 있을 때 그룹 안에서 생일이 같은 사람이 생길 확률은 ", np.longdouble(l[-1][0]), "이다. ")
     st.write("이론상 확률은 ", "$1 - { 366! \\over {366}^{%d} (366-%d)!}$" % (n, n),
              "≈", np.longdouble(f), "이다. ")
